# swagger_output.json
{
  "swagger":"2.0",
  "info":{
     "description":"\"Spring Boot REST API for telefônica project\"",
     "version":"1.0.0",
     "title":"Spring Boot REST API",
     "license":{
        "name":"Apache License Version 2.0",
        "url":"https://www.apache.org/licenses/LICENSE-2.0\""
     }
  },
  "host":"localhost:3000",
  "basePath":"/",
  "tags":[
     {
        "name":"pdf-controller",
        "description":"Pdf Controller"
     }
  ],
  "paths":{
     "/pdf/addImagem":{
        "post":{
           "tags":[
              "pdf-controller"
           ],
           "summary":"uploadFile",
           "operationId":"uploadFileUsingPOST",
           "consumes":[
              "multipart/form-data"
           ],
           "produces":[
              "*/*"
           ],
           "parameters":[
              {
                 "name":"pdf",
                 "in":"multipart/form-data",
                 "description":"pdf",
                 "required":true,
                 "type":"file"
              },
              {
                 "name":"imagem",
                 "in":"multipart/form-data",
                 "description":"png/jpeg/jpg",
                 "required":true,
                 "type":"file"
              },
              {
                 "name":"posicaoImagem",
                 "in":"query",
                 "description":"posicaoImagem",
                 "required":true,
                 "type":"string",
                 "enum":[
                    "ESQUEDA",
                    "CENTRO",
                    "DIREITA"
                 ]
              }
           ],
           "responses":{
              "200":{
                 "description":"OK",
                 "schema":{
                    "$ref":"#/definitions/Resource"
                 }
              },
              "201":{
                 "description":"Created"
              },
              "401":{
                 "description":"Unauthorized"
              },
              "403":{
                 "description":"Forbidden"
              },
              "404":{
                 "description":"Not Found"
              }
           },
           "deprecated":false
        }
     }
  },
  "definitions":{
     "File":{
        "type":"object",
        "properties":{
           "absolute":{
              "type":"boolean"
           },
           "absoluteFile":{
              "$ref":"#/definitions/File"
           },
           "absolutePath":{
              "type":"string"
           },
           "canonicalFile":{
              "$ref":"#/definitions/File"
           },
           "canonicalPath":{
              "type":"string"
           },
           "directory":{
              "type":"boolean"
           },
           "file":{
              "type":"boolean"
           },
           "freeSpace":{
              "type":"integer",
              "format":"int64"
           },
           "hidden":{
              "type":"boolean"
           },
           "name":{
              "type":"string"
           },
           "parent":{
              "type":"string"
           },
           "parentFile":{
              "$ref":"#/definitions/File"
           },
           "path":{
              "type":"string"
           },
           "totalSpace":{
              "type":"integer",
              "format":"int64"
           },
           "usableSpace":{
              "type":"integer",
              "format":"int64"
           }
        },
        "title":"File"
     },
     "InputStream":{
        "type":"object",
        "title":"InputStream"
     },
     "Resource":{
        "type":"object",
        "properties":{
           "description":{
              "type":"string"
           },
           "file":{
              "$ref":"#/definitions/File"
           },
           "filename":{
              "type":"string"
           },
           "inputStream":{
              "$ref":"#/definitions/InputStream"
           },
           "open":{
              "type":"boolean"
           },
           "readable":{
              "type":"boolean"
           },
           "uri":{
              "$ref":"#/definitions/URI"
           },
           "url":{
              "$ref":"#/definitions/URL"
           }
        },
        "title":"Resource"
     },
     "URI":{
        "type":"object",
        "properties":{
           "absolute":{
              "type":"boolean"
           },
           "authority":{
              "type":"string"
           },
           "fragment":{
              "type":"string"
           },
           "host":{
              "type":"string"
           },
           "opaque":{
              "type":"boolean"
           },
           "path":{
              "type":"string"
           },
           "port":{
              "type":"integer",
              "format":"int32"
           },
           "query":{
              "type":"string"
           },
           "rawAuthority":{
              "type":"string"
           },
           "rawFragment":{
              "type":"string"
           },
           "rawPath":{
              "type":"string"
           },
           "rawQuery":{
              "type":"string"
           },
           "rawSchemeSpecificPart":{
              "type":"string"
           },
           "rawUserInfo":{
              "type":"string"
           },
           "scheme":{
              "type":"string"
           },
           "schemeSpecificPart":{
              "type":"string"
           },
           "userInfo":{
              "type":"string"
           }
        },
        "title":"URI"
     },
     "URL":{
        "type":"object",
        "properties":{
           "authority":{
              "type":"string"
           },
           "content":{
              "type":"object"
           },
           "defaultPort":{
              "type":"integer",
              "format":"int32"
           },
           "file":{
              "type":"string"
           },
           "host":{
              "type":"string"
           },
           "path":{
              "type":"string"
           },
           "port":{
              "type":"integer",
              "format":"int32"
           },
           "protocol":{
              "type":"string"
           },
           "query":{
              "type":"string"
           },
           "ref":{
              "type":"string"
           },
           "userInfo":{
              "type":"string"
           }
        },
        "title":"URL"
     }
  }
}